+++ engine/docs/changelog.md
@@ -4,6 +4,7 @@
 * Removed cleanMetadataAcquisitionTestsuites from enclave_manager
 * Removed docker specific code from enclave_manager
 * Removed docker_network_allocator code
+* Updated enclave_manager to handle reporting on stopped containers where the GRPC port_specs are nil
 
 # 1.13.3
 ### Features
@@ -15,6 +16,7 @@
 * Upgraded to Kurt Core 1.41.2 which uses `files artifact expansion volumes` and `files artifact expander` objects
 
 # 1.13.2
+
 ### Fixes
 * Bumped dependencies for Core 1.41.1 and Container Engine Lib 0.13.0