+++ engine/docs/changelog.md
@@ -1,4 +1,13 @@
 # TBD
+### Breaking Changes
+* Updated enclave_manager to use kurtosis_backend for enclave operations
+* Updated enclave_manager to use kurtosis_backend for api container operations in an enclave
+* Removed cleanMetadataAcquisitionTestsuites from enclave_manager
+* Removed docker specific code from enclave_manager
+* Removed docker_network_allocator code
+
+### Fixes
+* Updated enclave_manager to handle reporting on stopped containers where the GRPC port_specs are nil
 
 # 1.13.3
 ### Features
@@ -10,6 +19,7 @@
 * Upgraded to Kurt Core 1.41.2 which uses `files artifact expansion volumes` and `files artifact expander` objects
 
 # 1.13.2
+
 ### Fixes
 * Bumped dependencies for Core 1.41.1 and Container Engine Lib 0.13.0