+++ engine/docs/changelog.md
@@ -1,4 +1,6 @@
 # TBD
+
+# 1.14.0
 ### Breaking Changes
 * Updated enclave_manager to use kurtosis_backend for enclave operations
 * Updated enclave_manager to use kurtosis_backend for api container operations in an enclave